+++ test/interactive/scene_controller_interactive_unit_test.dart
@@ -5,7 +5,6 @@ import 'package:iwb_canvas_engine/iwb_canvas_engine.dart';
 import 'package:iwb_canvas_engine/src/core/nodes.dart' hide NodeId;
 import 'package:iwb_canvas_engine/src/core/scene.dart';
 import 'package:iwb_canvas_engine/src/core/grid_safety_limits.dart';
-import 'package:iwb_canvas_engine/src/core/pointer_input.dart';
 import 'package:iwb_canvas_engine/src/model/document.dart';
 
 NodeSnapshot _nodeById(SceneSnapshot snapshot, NodeId id) {